Make `QuarkusLiquibaseConnectionProvider` extending `DefaultLiquibaseConnectionProvider` to allow `indexCreationThreshold` working
Closes #32860
