+++ quarkus/runtime/src/main/resources/META-INF/services/org.keycloak.connections.jpa.JpaConnectionProviderFactory
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.keycloak.quarkus.runtime.storage.legacy.database.QuarkusJpaConnectionProviderFactory
+org.keycloak.quarkus.runtime.storage.database.jpa.QuarkusJpaConnectionProviderFactory
